Support for rudimentary transclusion.
Syntax is:

{{tiddler title}}

on a line by itself.

At the moment this just looks in the bag where the containing
tiddler exists, however, it will eventually do:

* look in the recipe of the containing tiddler
* if not recipe look in the bag

And there will be:

{{tiddler title}}@space

to do interspace transclusion

Might make @fnd do all that though.

Note that this calls the upperlever wikitext_render so we can
transclude tiddlywiki text or anything else renderable if desired.

twikified.py (in twikifier) will need to be fixed similarly