Begin the process of switching over to using Python-Markdown
This gets us easier extensions from us, and access to useful
extensions that are built in, such as fenced code blocks (that is
github style ``` code blocks).

This is all working as well as the previous code, except for
transclusion, which has yet to be implemented, but is likely to
be implemented in much the same way are the prior lib.

Some important changes:

* plan uris are no longer processed, <http://fdafdsa> is the default
  way, so using that for now. I might fix that if someone like @FND complains
* https://github.com/waylan/Python-Markdown/issues/196
* links get a class of wikilink
* code is broken down into multiple modules to support the extension
  model

A bug I've just realized:

* if wiki_link_base is not set, none of the extension code gets
  added, this duplicates the old behavior but I'm not sure if that
  makes sense. Maybe the name is wrong. Perhaps just 'link_base',
  but then what about backwards compatibility

Todo:

* implement transclusion
* test the newly added (internal) extensions