Get transclusion working.
This is essentially the same code as before, just prepped
a bit differntly:

Transclusion is handled as postprocessor, so our regexp is looking
for a paragraph containing just {{title}} and then doing some
heuristics to determine the correct bag in which to find that tiddler.

As before we render the content using the higher renderer, so we can
mix markdown and other renderable types (such at tiddlywiki) just fine.