make inter-linking and -transclusion extensible
In previous version of the code, links with @foo only worked
for tiddlyspace. This change makes that handling extensible
via configuration:

* `markdown.interlinker` names a function which returns a link
  to whatever might be considered a space or target of some
  kind

* `markdown.transclude_url` names a function which returns a
  link to a tiddler in a target

* `markdown.target_resolver` names a function which figures out
  the tiddler object that is to be transcluded, when a
  {{foobar}}@target type transclusion is being used

See the tests that set those things for some understanding of
what they do, how they work.

I'd appreciate some review on this from @colmjude and @FND for
two reasons:

* this changes functionality in a usefully extensible way that might
  work okay for things like BFW and other stuff
* if this code is installed in a @tiddlyspace environment the
  tiddlywebplugins.tiddlyspace.config file will need some small
  updates (to reflect the tests here, basically)

In case it's not clear, I'm doing this to enable inter-linking between
"tanks".